Changed a few things about variables

+++ src/main.cpp
@@ -15,7 +15,7 @@ uint8_t ledState = LOW; //"ledState" is used to store the inbuilt LED's state
 unsigned long debounceDelay = 100;  //Debounce time interval (button ignore time)
 unsigned long lastDebounceTime = 0; //This will store the last time the button was pressed
 bool allowDispVal = true; //Boolean to decide to allow display code execution
-int32_t frequency = 1000; //Initialise the PWM frequency
+int16_t frequency = 1000; //Initialise the PWM frequency
 double outPWM = 50*655; //Initialise (in this case, set) the PWM duty cycle set to 50%
 
 
@@ -46,7 +46,7 @@ void setup() {
 
 
 void loop() {
-  //Assigns the potentiometer position to a variable
+  //Assigns the 10-bit potentiometer position to a variable
   frequency = map(analogRead(potInVal), 0, 903, 10, 5000);
 
   SetPinFrequencySafe(signal_out_pin, frequency); //Live set/edit PWM frequency
@@ -57,8 +57,7 @@ void loop() {
     allowDispVal = true; //Set the serial display boolean to true
     digitalWrite(ledPin, LOW); //Turn the inbuilt led off
   }
-  else
-  {
+  else  {
     allowDispVal = false; //Set the serial display boolean to false
     digitalWrite(ledPin, HIGH); //Turn the inbuilt led on
   }
